--- frontend/src/components/Header.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const Header = () => {
-  return (
-    <header className="app-header">
-      <img
-        src=""
-        alt="Finance Companion Logo"
-      />
-      <h1 className="header-title">Finance Companion</h1>
-      <button className="sign-out-btn">Sign Out</button>
-    </header>
-  );
-};
-
-export default Header;
