+++ frontend/src/context/ProtectedRoute.jsx
@@ -1,6 +1,6 @@
-import Loading from "./Loading";
+import Loading from "../shared-components/Loading";
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "./AuthContext";
 
 const ProtectedRoute = ({ children }) => {
   const { user, loading } = useAuth();
